requirements.txt
langchain
langchain-openai
openai
python-dotenv
pydantic
rich
guardrails-ai
pydantic
typing_extensions
ftfy
promptfoo
langchain_community
# API
fastapi>=0.111.0
uvicorn>=0.30.0

# Tests & client
pytest>=8.0.0
httpx>=0.27.0